feat: add Vercel Edge API for dynamic Open Graph image generation and deprecate legacy SST endpoint
